Combined name_disambiguation and name_preprocessing
Put them into the name_preprocessing doc
